Add UTs for UpdateInputWithUsernameFromIdToken
